jaxlib 0.4.36

+++ recipe/build.sh
@@ -4,6 +4,8 @@ set -euxo pipefail
 
 export JAX_RELEASE=$PKG_VERSION
 
+$RECIPE_DIR/add_py_toolchain.sh
+
 if [[ "${target_platform}" == osx-* ]]; then
   export LDFLAGS="${LDFLAGS} -lz -framework CoreFoundation -Xlinker -undefined -Xlinker dynamic_lookup"
   # Remove stdlib=libc++; this is the default and errors on C sources.
@@ -15,8 +17,6 @@ export CFLAGS="${CFLAGS} -DNDEBUG"
 export CXXFLAGS="${CXXFLAGS} -DNDEBUG"
 
 if [[ "${cuda_compiler_version:-None}" != "None" ]]; then
-    # Remove incompatible argument from bazelrc
-    sed -i '/Qunused-arguments/d' .bazelrc
     if [[ ${cuda_compiler_version} == 11.8 ]]; then
         export HERMETIC_CUDA_COMPUTE_CAPABILITIES=sm_35,sm_50,sm_60,sm_62,sm_70,sm_72,sm_75,sm_80,sm_86,sm_87,sm_89,sm_90,compute_90
         export TF_CUDA_PATHS="${CUDA_HOME},${PREFIX}"
@@ -42,9 +42,9 @@ if [[ "${cuda_compiler_version:-None}" != "None" ]]; then
         export LOCAL_CUDNN_PATH="${PREFIX}/targets/x86_64-linux"
         export LOCAL_NCCL_PATH="${PREFIX}/targets/x86_64-linux"
         mkdir -p ${BUILD_PREFIX}/targets/x86_64-linux/bin
-        ln -s $(which ptxas) ${BUILD_PREFIX}/targets/x86_64-linux/bin/ptxas
-        ln -s $(which nvlink) ${BUILD_PREFIX}/targets/x86_64-linux/bin/nvlink
-        ln -s $(which fatbinary) ${BUILD_PREFIX}/targets/x86_64-linux/bin/fatbinary
+        test -f ${BUILD_PREFIX}/targets/x86_64-linux/bin/ptxas || ln -s $(which ptxas) ${BUILD_PREFIX}/targets/x86_64-linux/bin/ptxas
+        test -f ${BUILD_PREFIX}/targets/x86_64-linux/bin/nvlink || ln -s $(which nvlink) ${BUILD_PREFIX}/targets/x86_64-linux/bin/nvlink
+        test -f ${BUILD_PREFIX}/targets/x86_64-linux/bin/fatbinary || ln -s $(which fatbinary) ${BUILD_PREFIX}/targets/x86_64-linux/bin/fatbinary
     else
         echo "unsupported cuda version."
         exit 1
@@ -58,9 +58,9 @@ if [[ "${cuda_compiler_version:-None}" != "None" ]]; then
     export TF_NEED_CUDA=1
     export TF_NCCL_VERSION=$(pkg-config nccl --modversion | grep -Po '\d+\.\d+')
     export CUDA_COMPILER_MAJOR_VERSION=$(echo "$cuda_compiler_version" | cut -d '.' -f 1)
-    CUDA_ARGS="--enable_cuda --build_gpu_plugin --gpu_plugin_cuda_version=${CUDA_COMPILER_MAJOR_VERSION} \
-               --enable_nccl \
+    CUDA_ARGS="--wheels=jaxlib,jax-cuda-plugin,jax-cuda-pjrt \
                --cuda_compute_capabilities=$HERMETIC_CUDA_COMPUTE_CAPABILITIES \
+               --cuda_major_version=${CUDA_COMPILER_MAJOR_VERSION} \
                --cuda_version=$TF_CUDA_VERSION \
                --cudnn_version=$TF_CUDNN_VERSION"
 fi
@@ -74,6 +74,10 @@ source gen-bazel-toolchain
 cat >> .bazelrc <<EOF
 
 build --crosstool_top=//bazel_toolchain:toolchain
+build --platforms=//bazel_toolchain:target_platform
+build --host_platform=//bazel_toolchain:build_platform
+build --extra_toolchains=//bazel_toolchain:cc_cf_toolchain
+build --extra_toolchains=//bazel_toolchain:cc_cf_host_toolchain
 build --logging=6
 build --verbose_failures
 build --toolchain_resolution_debug
@@ -81,6 +85,9 @@ build --define=PREFIX=${PREFIX}
 build --define=PROTOBUF_INCLUDE_PATH=${PREFIX}/include
 build --local_cpu_resources=${CPU_COUNT}
 build --define=with_cross_compiler_support=true
+
+# We need to define a dummy value for this as we delete everything else for build_cuda_with_nvcc
+build:build_cuda_with_nvcc --action_env=CONDA_USE_NVCC=1
 EOF
 
 if [[ "${target_platform}" == "osx-arm64" || "${target_platform}" != "${build_platform}" ]]; then
@@ -107,11 +114,18 @@ else
     EXTRA="${CUDA_ARGS:-}"
 fi
 if [[ "${target_platform}" == linux-* ]]; then
-    EXTRA="${EXTRA} --nouse_clang"
+    EXTRA="${EXTRA} --use_clang false"
+
+    # Remove incompatible argument from bazelrc
+    sed -i '/Qunused-arguments/d' .bazelrc
+    # Don't override our toolchain for CUDA
+    sed -i '/TF_NVCC_CLANG/{N;d}' .bazelrc
+    # Keep using our toolchain
+    sed -i '/--crosstool_top=@local_config_cuda/d' .bazelrc
 fi
-${PYTHON} build/build.py \
+
+${PYTHON} build/build.py build \
     --target_cpu_features default \
-    --enable_mkl_dnn \
     ${EXTRA}
 
 # Clean up to speedup postprocessing
